Made dependency on nanomsg conditional
This greatly limits the number of supported features, both core features
such as learning and ageing, and tooling such as the event logger and
the debugger. However other packet forwarding capabilities should be
unaffected.

A new configure flag has been introduced, '--without-nanomsg', to build
without nanomsg.
