+++ src/bm_sim/Makefile.am
@@ -62,6 +62,7 @@ simple_pre_lag.cpp \
 tables.cpp \
 target_parser.cpp \
 transport.cpp \
+transport_nn.cpp \
 utils.h \
 version.cpp \
 version.h \
